feat(familiars): click-to-upload avatar + images-above-icons in Look tab
Two Familiar Studio improvements:

1. The header avatar is now a click-to-upload control — clicking it (or
   keyboard-activating) opens the native file picker and sets a new avatar
   image for the familiar. A camera overlay fades in on hover/focus to signal
   it's interactive; the avatar updates reactively from the image store.

2. The Look tab now orders the "Avatar image" section above "Icon" (was
   Icon → color → image), so picking a photo is the first choice.

Image-upload logic (prepare/downsize + commit) is extracted into a shared
`useFamiliarImageUpload` hook (src/lib/familiar-image-upload.ts) reused by both
the Look-tab dropzone and the header avatar, removing the duplicate.
