+++ src/app/globals.css
@@ -4354,6 +4354,14 @@ button:active > .edge-rail-chip {
 .familiar-studio__main { display: grid; grid-template-rows: auto 1fr auto; min-height: 0; }
 .familiar-studio__header { display: flex; align-items: center; gap: var(--space-3); padding: var(--space-4) var(--space-5); border-bottom: 1px solid var(--border-hairline); }
 .familiar-studio__heading { display: flex; flex-direction: column; gap: 1px; min-width: 0; flex: 1; }
+/* Header avatar = click-to-upload. The avatar fills the button; a camera
+   overlay fades in on hover/focus to signal it's interactive. */
+.familiar-studio__avatar-wrap { position: relative; display: inline-flex; flex-shrink: 0; }
+.familiar-studio__avatar-btn { position: relative; display: grid; place-items: center; padding: 0; border: none; background: none; cursor: pointer; border-radius: var(--radius-md); }
+.familiar-studio__avatar-overlay { position: absolute; inset: 0; display: grid; place-items: center; border-radius: inherit; color: #fff; background: color-mix(in oklch, #000 52%, transparent); opacity: 0; transition: opacity var(--duration-fast) var(--ease-standard); }
+.familiar-studio__avatar-btn:hover .familiar-studio__avatar-overlay,
+.familiar-studio__avatar-btn:focus-visible .familiar-studio__avatar-overlay { opacity: 1; }
+.familiar-studio__avatar-toast { position: absolute; top: calc(100% + 4px); left: 0; white-space: nowrap; font-size: var(--text-xs); color: var(--text-muted); }
 .familiar-studio__name { font-size: var(--text-lg); font-weight: 600; letter-spacing: -0.01em; color: var(--text-primary); background: transparent; border: none; padding: 0; text-align: left; cursor: text; font-family: inherit; min-width: 0; border-radius: var(--radius-sm); transition: background var(--duration-fast) var(--ease-standard); }
 .familiar-studio__name:hover:not(.familiar-studio__name--editing) { background: var(--bg-raised); padding: 0 5px; margin-left: -5px; }
 .familiar-studio__name--editing { background: var(--bg-raised); border: 1px solid var(--border-strong); border-radius: var(--radius-sm); padding: 2px 6px; outline: none; box-shadow: 0 0 0 3px var(--ring-focus-soft); }
